projeto/backend/requirements.txt
# Com carinho na base tecnica, estas dependencias sustentam API, JWT, MongoDB e validacoes.
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-dotenv